Added validation with zod to the requests body and params

+++ src/modules/patient/patient.controller.ts
@@ -1,21 +1,30 @@
 import { FastifyReply, FastifyRequest } from "fastify";
 import prisma from "../../utils/prisma";
-import { ICreatePatient, IParams, IUpdatePatient } from "./patient.schema";
+import {
+  CreatePatientRequestBodySchema,
+  GetPatientByIdRequestParamsSchema,
+  PutPatientByIdRequestBodySchema,
+  CreatePatientRequestBodyType,
+  GetPatientByIdRequestParamsType,
+  PutPatientByIdRequestBodyType,
+} from "./patient.schema";
 
 export async function createPatientHandler(
-  request: FastifyRequest<{ Querystring: ICreatePatient }>,
+  request: FastifyRequest<{ Body: CreatePatientRequestBodyType }>,
   reply: FastifyReply
 ) {
-  const { name, email } = request.query;
+  const parsedBody = CreatePatientRequestBodySchema.safeParse(request.body);
 
-  const patient = await prisma.patient.create({
-    data: {
-      name,
-      email,
-    },
-  });
-
-  reply.send(patient);
+  if (parsedBody.success) {
+    try {
+      const patient = await prisma.patient.create({ data: parsedBody.data });
+      reply.send(patient);
+    } catch (err) {
+      reply.send(err);
+    }
+  } else {
+    reply.code(500).send(parsedBody.error);
+  }
 }
 
 export async function getAllPatientsHandler(
@@ -28,45 +37,72 @@ export async function getAllPatientsHandler(
 }
 
 export async function getPatientByIdHandler(
-  request: FastifyRequest<{ Params: IParams }>,
+  request: FastifyRequest<{ Params: GetPatientByIdRequestParamsType }>,
   reply: FastifyReply
 ) {
-  const { patientId } = request.params;
-  const patient = await prisma.patient.findUnique({
-    where: { id: Number(patientId) },
-  });
-  reply.send(patient);
+  const parsedParams = GetPatientByIdRequestParamsSchema.safeParse(
+    request.params
+  );
+
+  if (parsedParams.success) {
+    try {
+      const patient = await prisma.patient.findUnique({
+        where: { id: parsedParams.data.patientId },
+      });
+
+      reply.send(patient);
+    } catch (err) {
+      reply.send(err);
+    }
+  } else {
+    reply.code(500).send(parsedParams.error);
+  }
 }
 
 export async function updatePatientByIdHandler(
-  request: FastifyRequest<{ Params: IParams; Querystring: IUpdatePatient }>,
+  request: FastifyRequest<{
+    Params: GetPatientByIdRequestParamsType;
+    Body: PutPatientByIdRequestBodyType;
+  }>,
   reply: FastifyReply
 ) {
-  const { patientId } = request.params;
-  const { name, email } = request.query;
-
-  const data = {
-    ...(name && { name }),
-    ...(email && { email }),
-  };
-
-  const patient = await prisma.patient.update({
-    where: { id: Number(patientId) },
-    data,
-  });
-
-  reply.send(patient);
+  const parsedParams = GetPatientByIdRequestParamsSchema.safeParse(
+    request.params
+  );
+  const parsedBody = PutPatientByIdRequestBodySchema.safeParse(request.body);
+  console.log(parsedBody, parsedParams);
+  if (parsedBody.success && parsedParams.success) {
+    try {
+      const patient = await prisma.patient.update({
+        where: { id: parsedParams.data.patientId },
+        data: parsedBody.data,
+      });
+      reply.send(patient);
+    } catch (err) {
+      reply.send(err);
+    }
+  } else if (!parsedBody.success) reply.send(parsedBody.error);
+  else if (!parsedParams.success) reply.send(parsedParams.error);
 }
 
 export async function deletePatientByIdHandler(
-  request: FastifyRequest<{ Params: IParams }>,
+  request: FastifyRequest<{ Params: GetPatientByIdRequestParamsType }>,
   reply: FastifyReply
 ) {
-  const { patientId } = request.params;
+  const parsedParams = GetPatientByIdRequestParamsSchema.safeParse(
+    request.params
+  );
 
-  const patient = await prisma.patient.delete({
-    where: { id: Number(patientId) },
-  });
-
-  reply.send(patient);
+  if (parsedParams.success) {
+    try {
+      const patient = await prisma.patient.delete({
+        where: { id: parsedParams.data.patientId },
+      });
+      reply.send(patient);
+    } catch (err) {
+      reply.send(err);
+    }
+  } else {
+    reply.send(parsedParams.error);
+  }
 }
